Fix: fixtures loading with single inheritance
I wasn't able to load fixtures with inheritance because when I
defined a Novel in my fixtures (which extends Book), it would be saved
by the sfPropelData class in a 'object_references' property under the
'%class%_%pk%' key (Novel_42 for instance).

Unfortunately, if I reference later in my fixtures this novel, the class
would look up in the cache but with the wrong key (it would use Book_42).